requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.82.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
orjson==3.8.0
psycopg2-binary==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
